+++ contracts/CelerToken.sol
@@ -6,8 +6,8 @@ import "./lib/external/openzeppelin-solidity/contracts/token/ERC20/PausableToken
 contract CelerToken is PausableToken {
   string public name = "CelerToken";
   string public symbol = "CELR";
-  uint public decimals = 18;
-  uint public INITIAL_SUPPLY = (10 ** 10) * (10 ** decimals);
+  uint256 public decimals = 18;
+  uint256 public INITIAL_SUPPLY = (10 ** 10) * (10 ** decimals);
 
   constructor() public {
     totalSupply_ = INITIAL_SUPPLY;
